Move files around to crc package
