+++ plugins/org.jboss.tools.openshift.cdk.server/plugin.properties
@@ -20,5 +20,5 @@ launch_cdk_server=Launch Container Development Environment using Red Hat CDK
 cdk.detector.2x=Container Development Environment 2.x
 cdk.detector.3=Container Development Environment 3.x
 
-openshift.crc.100.name=Red Hat CodeReady Containers 1.0
+openshift.crc.100.name=Red Hat CodeReady Containers 1.0 (Tech Preview)
 openshift.crc.100.description=Integration and support for Red Hat CodeReady Containers 1.0
