+++ plugins/org.jboss.tools.openshift.cdk.server/src/org/jboss/tools/openshift/internal/cdk/server/core/adapter/controllers/AbstractCDKLaunchController.java
@@ -84,34 +84,16 @@ public abstract class AbstractCDKLaunchController extends AbstractSubsystemContr
 	}
 
 	protected IProcess addProcessToLaunch(Process p, ILaunch launch, IServer s, boolean terminal, String cmdLoc) {
-		Map<String, String> processAttributes = new HashMap<String, String>();
-		String progName = new Path(cmdLoc).lastSegment();
-		if (terminal) {
-			launch.setAttribute(DebugPlugin.ATTR_CAPTURE_OUTPUT, "false");
-		}
-		processAttributes.put(IProcess.ATTR_PROCESS_TYPE, progName);
-		IProcess process = createProcess(terminal, launch, p, cmdLoc, processAttributes);
-		launch.addProcess(process);
-
+		IProcess process = ProcessUtility.addProcessToLaunch(p, launch, s, terminal, cmdLoc);
 		if (terminal) {
 			linkTerminal(p);
 		}
 		return process;
 	}
 
-	protected IProcess createProcess(boolean terminal, ILaunch launch, Process p, String cmd,
-			Map<String, String> attr) {
-		IProcess process = null;
-		if (terminal) {
-			process = new RuntimeProcess(launch, p, cmd, attr) {
-				protected IStreamsProxy createStreamsProxy() {
-					return null;
-				}
-			};
-		} else {
-			process = new RuntimeProcess(launch, p, cmd, attr);
-		}
-		return process;
+	protected IProcess createProcess(boolean terminal, ILaunch launch, Process p, 
+			String cmd, Map<String, String> attr) {
+		return ProcessUtility.createProcess(terminal, launch, p, cmd, attr);
 	}
 
 	protected void linkTerminal(Process p) {
