+++ plugins/org.jboss.tools.openshift.cdk.server/src/org/jboss/tools/openshift/internal/crc/server/core/adapter/CRC100Server.java
@@ -8,7 +8,7 @@
  * Contributors: 
  * Red Hat, Inc. - initial API and implementation 
  ******************************************************************************/
-package org.jboss.tools.openshift.internal.cdk.server.core.adapter;
+package org.jboss.tools.openshift.internal.crc.server.core.adapter;
 
 import java.io.File;
 
@@ -23,7 +23,7 @@ import org.jboss.tools.openshift.common.core.utils.StringUtils;
 import org.jboss.tools.openshift.internal.cdk.server.core.BinaryUtility;
 
 public class CRC100Server extends ServerDelegate {
-	public static final String CRC_100_SERVER_TYPE_ID = "org.jboss.tools.openshift.cdk.server.type.crc.v100";
+	public static final String CRC_100_SERVER_TYPE_ID = "org.jboss.tools.openshift.crc.server.type.crc.v100";
 	private static final String CRC_10_BASE_NAME = "CodeReady Containers 1.0";
 	public static final String PROPERTY_PULL_SECRET_FILE = "crc.pullsecret.file";
 	public static final String PROPERTY_BINARY_FILE = "crc.binary.file";
